Knowledge base.
Knowledge base system created with pure HTML, CSS, and Javascript.

In this system, it's possible to create new cards with relevant information and edit existing cards.

Because there's no HTTP server, after creating the cards it's necessary to use the Export button, which will generate the new Items and Categories class in a JSON file. Copy the file's contents to main.js, replacing the initial class data.